+++ app/routes/signup.tsx
@@ -1,46 +1,58 @@
 import React, { useState } from "react";
-import { Layout } from "../../components/layout";
-import { Textfield } from "~/components/textField";
 import { Link, useActionData } from "@remix-run/react";
 import { MetaFunction } from "@remix-run/node";
 
+import { Layout } from "../components/layout";
+import { Textfield } from "../components/textField";
+
+import { loader, action } from "../server/signup.server";
+
 export const meta: MetaFunction = () => {
-  return [{ title: "New Remix App Login" }];
+  return [{ title: "New Remix App Signup" }];
 };
 
 interface ActionData {
   fields?: {
     email?: string;
     password?: string;
+    name?: string;
   };
   error?: string;
 }
 
-export default function Login() {
+export default function SignUp() {
   const actionData = useActionData<ActionData>();
   const [formData, setFormData] = useState({
     email: actionData?.fields?.email || "",
     password: actionData?.fields?.password || "",
+    name: actionData?.fields?.name || "",
   });
 
-  // Updates the form data when an input changes
   const handleInputChange = (
     event: React.ChangeEvent<HTMLInputElement>,
     field: string
   ) => {
+    event.preventDefault();
     setFormData((form) => ({ ...form, [field]: event.target.value }));
   };
 
   return (
     <Layout>
       <div className="h-full justify-center bg-sky-400 items-center flex flex-col gap-y-5">
-        <form method="POST" className="rounded-lg bg-white p-6 w-96">
-          <h2 className="text-3xl font-bold text-gray-400 mb-5 flex flex-col items-center">
-            Login
+        <form method="POST" className="rounded-lg bg-white p-5 w-96">
+          <h2 className="text-3xl font-bold text-gray-400 mb-2 flex flex-col items-center">
+            Create an account
           </h2>
           {actionData?.error && (
             <p className="text-left ml-1 text-red-500">{actionData.error}</p>
           )}
+          <Textfield
+            htmlFor="name"
+            type="name"
+            placeholder="Name"
+            value={formData.name}
+            onChange={(e) => handleInputChange(e, "name")}
+          />
           <Textfield
             htmlFor="email"
             placeholder="Email"
@@ -54,21 +66,21 @@ export default function Login() {
             value={formData.password}
             onChange={(e) => handleInputChange(e, "password")}
           />
-          <div className="w-full text-center mt-5">
+          <div className="w-full text-center mt-2">
             <button
               type="submit"
               name="_action"
               value="Sign In"
-              className="w-full rounded-xl mt-2 bg-sky-400 px-3 py-2 text-white font-semibold transition duration-300 ease-in-out hover:bg-sky-600"
+              className="w-full rounded-xl bg-sky-400 px-3 py-2 text-white font-semibold transition duration-300 ease-in-out hover:bg-sky-600"
             >
-              Login
+              Create an account
             </button>
           </div>
         </form>
         <p className="text-white">
-          Don't have an account?{" "}
-          <Link to="/signup">
-            <span className="text-white px-2 underline">Signup</span>
+          Already have an account?
+          <Link to="/login">
+            <span className="text-white px-2 underline">Sign In</span>
           </Link>
         </p>
       </div>
@@ -76,4 +88,4 @@ export default function Login() {
   );
 }
 
-export { loader, action } from "./login.server";
+export { loader, action };
